requirements.txt
scikit-learn
matplotlib
seaborn
nltk
string
re
wordcloud
collection
xgboost
sklearn.feature_extraction.text